+++ internal/plugin/service/vpc/projectvpc/projectvpc_acc_test.go
@@ -1,13 +1,14 @@
-package vpc_test
+package projectvpc_test
 
 import (
 	"context"
 	"errors"
 	"fmt"
+	"net/http"
 	"regexp"
 	"testing"
 
-	"github.com/aiven/aiven-go-client/v2"
+	avngen "github.com/aiven/go-client-codegen"
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
@@ -42,6 +43,7 @@ func TestAccAivenProjectVPC_basic(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "cloud_name", "google-europe-west2"),
 					resource.TestCheckResourceAttr(resourceName, "network_cidr", "192.168.0.0/24"),
 					resource.TestCheckResourceAttr(resourceName, "state", "ACTIVE"),
+					resource.TestCheckResourceAttrSet(resourceName, "project_vpc_id"),
 				),
 			},
 			{
@@ -52,7 +54,49 @@ func TestAccAivenProjectVPC_basic(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "cloud_name", "azure-westeurope"),
 					resource.TestCheckResourceAttr(resourceName, "network_cidr", "192.168.1.0/24"),
 					resource.TestCheckResourceAttr(resourceName, "state", "ACTIVE"),
+					resource.TestCheckResourceAttrSet(resourceName, "project_vpc_id"),
 				),
+			},
+			{
+				Config:            testAccProjectVPCResourceGetByID(),
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+		},
+	})
+}
+
+func TestAccAivenProjectVPC_backwardCompat(t *testing.T) {
+	resourceName := "aiven_project_vpc.bar"
+	project := acc.ProjectName()
+	config := testAccProjectVPCResourceOnly(project, "google-europe-west4", "192.168.2.0/24")
+	oldChecks := resource.ComposeTestCheckFunc(
+		resource.TestCheckResourceAttr(resourceName, "project", project),
+		resource.TestCheckResourceAttr(resourceName, "cloud_name", "google-europe-west4"),
+		resource.TestCheckResourceAttr(resourceName, "network_cidr", "192.168.2.0/24"),
+		resource.TestCheckResourceAttr(resourceName, "state", "ACTIVE"),
+	)
+	newChecks := resource.ComposeTestCheckFunc(
+		oldChecks,
+		resource.TestCheckResourceAttrSet(resourceName, "project_vpc_id"),
+	)
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck: func() { acc.TestAccPreCheck(t) },
+		Steps: []resource.TestStep{
+			{
+				ExternalProviders: map[string]resource.ExternalProvider{
+					"aiven": acc.ExternalAivenProvider(t, "4.56.0"),
+				},
+				Config: config,
+				Check:  oldChecks,
+			},
+			{
+				ProtoV6ProviderFactories: acc.TestProtoV6ProviderFactories,
+				Config:                   config,
+				ExpectNonEmptyPlan:       false,
+				Check:                    newChecks,
 			},
 		},
 	})
@@ -74,6 +118,16 @@ data "aiven_project_vpc" "vpc" {
   project    = aiven_project_vpc.bar.project
   cloud_name = aiven_project_vpc.bar.cloud_name
 }`, acc.ProjectName())
+}
+
+func testAccProjectVPCResourceOnly(project, cloud, cidr string) string {
+	return fmt.Sprintf(`
+resource "aiven_project_vpc" "bar" {
+  project      = %[1]q
+  cloud_name   = %[2]q
+  network_cidr = %[3]q
+}
+`, project, cloud, cidr)
 }
 
 func testAccCheckAivenProjectVPCAttributes(n string) resource.TestCheckFunc {
@@ -106,31 +160,32 @@ func testAccCheckAivenProjectVPCAttributes(n string) resource.TestCheckFunc {
 }
 
 func testAccCheckAivenProjectVPCResourceDestroy(s *terraform.State) error {
-	c := acc.GetTestAivenClient()
+	c, err := acc.GetTestGenAivenClient()
+	if err != nil {
+		return err
+	}
 
 	ctx := context.Background()
-
-	// loop through the resources in state, verifying each project VPC is destroyed
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "aiven_project_vpc" {
 			continue
 		}
 
-		projectName, vpcID, err := schemautil.SplitResourceID2(rs.Primary.ID)
+		project, vpcID, err := schemautil.SplitResourceID2(rs.Primary.ID)
 		if err != nil {
 			return err
 		}
 
-		vpc, err := c.VPCs.Get(ctx, projectName, vpcID)
+		vpc, err := c.VpcGet(ctx, project, vpcID)
 		if err != nil {
-			var e aiven.Error
-			if errors.As(err, &e) && e.Status != 404 && e.Status != 403 {
+			var e avngen.Error
+			if errors.As(err, &e) && e.Status != http.StatusNotFound && e.Status != http.StatusForbidden {
 				return err
 			}
 		}
 
 		if vpc != nil {
-			return fmt.Errorf("porject vpc (%s) still exists", rs.Primary.ID)
+			return fmt.Errorf("project vpc (%s) still exists", rs.Primary.ID)
 		}
 	}
 
@@ -175,105 +230,4 @@ resource "aiven_project_vpc" "bar" {
 data "aiven_project_vpc" "vpc2" {
   vpc_id = aiven_project_vpc.bar.id
 }`, acc.ProjectName())
-}
-
-// TestAccAivenProjectVPC_can_be_removed tests that `project_vpc_id` can be removed from a service
-// https://github.com/aiven/terraform-provider-aiven/issues/2242
-func TestAccAivenProjectVPC_can_be_removed(t *testing.T) {
-	ctx := context.Background()
-	project := acc.ProjectName()
-	resourceName := "aiven_grafana.foo"
-	serviceName := "test-acc-" + acc.RandStr()
-	cloudName := "google-europe-west12"
-	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:                 func() { acc.TestAccPreCheck(t) },
-		ProtoV6ProviderFactories: acc.TestProtoV6ProviderFactories,
-		CheckDestroy:             acc.TestAccCheckAivenServiceResourceDestroy,
-		Steps: []resource.TestStep{
-			{
-				// Creates a service with a VPC
-				Config: fmt.Sprintf(`
-resource "aiven_project_vpc" "foo" {
-  project      = %[1]q
-  cloud_name   = %[2]q
-  network_cidr = "192.168.1.0/24"
-}
-
-resource "aiven_grafana" "foo" {
-  project                 = aiven_project_vpc.foo.project
-  cloud_name              = aiven_project_vpc.foo.cloud_name
-  project_vpc_id          = aiven_project_vpc.foo.id
-  plan                    = "startup-1"
-  service_name            = %[3]q
-  maintenance_window_dow  = "monday"
-  maintenance_window_time = "10:00:00"
-}
-`, project, cloudName, serviceName),
-				Check: resource.ComposeTestCheckFunc(
-					func(s *terraform.State) error {
-						client, err := acc.GetTestGenAivenClient()
-						if err != nil {
-							return err
-						}
-
-						service, err := client.ServiceGet(ctx, project, serviceName)
-						if err != nil {
-							return err
-						}
-
-						// Should have a project VPC ID
-						if service.ProjectVpcId == "" {
-							return fmt.Errorf("expected service %q to have a project VPC ID", serviceName)
-						}
-
-						// Check that the project VPC ID is set
-						projectVPC := schemautil.BuildResourceID(project, service.ProjectVpcId)
-						return resource.TestCheckResourceAttr(resourceName, "project_vpc_id", projectVPC)(s)
-					},
-				),
-			},
-			{
-				// Removes the `project_vpc_id` field from the service
-				Config: fmt.Sprintf(`
-resource "aiven_project_vpc" "foo" {
-  project      = %[1]q
-  cloud_name   = %[2]q
-  network_cidr = "192.168.1.0/24"
-}
-
-resource "aiven_grafana" "foo" {
-  project                 = aiven_project_vpc.foo.project
-  cloud_name              = aiven_project_vpc.foo.cloud_name
-  plan                    = "startup-1"
-  service_name            = %[3]q
-  maintenance_window_dow  = "monday"
-  maintenance_window_time = "10:00:00"
-}
-`, project, cloudName, serviceName),
-				Check: resource.ComposeTestCheckFunc(
-					func(s *terraform.State) error {
-						client, err := acc.GetTestGenAivenClient()
-						if err != nil {
-							return err
-						}
-
-						service, err := client.ServiceGet(ctx, project, serviceName)
-						if err != nil {
-							return err
-						}
-
-						// Must be empty
-						if service.ProjectVpcId != "" {
-							return fmt.Errorf("expected service %q to not have a project VPC ID", serviceName)
-						}
-
-						// Not set in TF either (not computed).
-						// Terraform can't distinguish between "not set" and "empty string",
-						// so we check for an empty string here.
-						return resource.TestCheckResourceAttr(resourceName, "project_vpc_id", "")(s)
-					},
-				),
-			},
-		},
-	})
 }